+++ source/PC/numberDetection/train_number_card.py
@@ -8,27 +8,10 @@ from chainer import training
 from chainer.training import extensions
 import chainerx
 from chainer.datasets import LabeledImageDataset
+from MLP import MLP
 
 import matplotlib
 matplotlib.use('Agg')
-
-
-# Network definition
-class MLP(chainer.Chain):
-
-    def __init__(self, n_units, n_out):
-        super(MLP, self).__init__()
-        with self.init_scope():
-            # the size of the inputs to each layer will be inferred
-            self.l1 = L.Linear(None, n_units)  # n_in -> n_units
-            self.l2 = L.Linear(None, n_units)  # n_units -> n_units
-            self.l3 = L.Linear(None, n_out)  # n_units -> n_out
-
-    def forward(self, x):
-        h1 = F.relu(self.l1(x))
-        h2 = F.relu(self.l2(h1))
-        return self.l3(h2)
-
 
 def main():
     parser = argparse.ArgumentParser(description='Chainer example: MNIST')
